+++ packages/sources/cfbenchmarks/test/integration/fixtures.ts
@@ -28,6 +28,61 @@ export const mockResponseSuccess = (): nock.Scope =>
       ],
     )
 
+export const mockBircResponseSuccess = (): nock.Scope =>
+  nock('https://www.cfbenchmarks.com/api', {
+    reqheaders: {
+      Authorization: 'Basic ZmFrZS1hcGktdXNlcm5hbWU6ZmFrZS1hcGktcGFzc3dvcmQ=',
+    },
+  })
+    .get('/v1/curves')
+    .query({ id: 'BIRC' })
+    .reply(
+      200,
+      {
+        serverTime: '2023-02-24T08:17:17.446Z',
+        payload: [
+          {
+            tenors: {
+              SIRB: '0.0986',
+              '1W': '0.0077',
+              '2W': '0.0186',
+              '3W': '0.0219',
+              '1M': '0.0168',
+              '2M': '0.0099',
+              '3M': '0.0097',
+              '4M': '0.0078',
+              '5M': '0.0059',
+            },
+            time: 1659366000000,
+          },
+          {
+            tenors: {
+              SIRB: '0.0947',
+              '1W': '0.0367',
+              '2W': '0.0185',
+              '3W': '0.0229',
+              '1M': '0.0274',
+              '2M': '0.0297',
+              '3M': '0.0275',
+              '4M': '0.0253',
+              '5M': '0.0000',
+            },
+            time: 1677168000000,
+          },
+        ],
+      },
+      [
+        'Content-Type',
+        'application/json',
+        'Connection',
+        'close',
+        'Vary',
+        'Accept-Encoding',
+        'Vary',
+        'Origin',
+      ],
+    )
+
 export const mockSubscribeResponse = {
   request: {
     type: 'subscribe',